requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==24.2.0
beir==2.0.0
chardet==5.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
datasets==2.21.0
dill==0.3.8
distro==1.9.0
elasticsearch==7.9.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-ollama==0.1.3
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.116
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
ollama==0.3.2
openai==1.44.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.2
pyarrow==17.0.0
pydantic==2.9.0
pydantic_core==2.23.2
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrec_eval_terrier==0.5.6
pytz==2024.1
PyYAML==6.0.2
ragas==0.1.16
regex==2024.7.24
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
wheel==0.43.0
xxhash==3.5.0
yarl==1.10.0
